refactor: Change project structure.
