+++ src/main/java/com/example/welcomeclient/services/CallServiceImpl.java
@@ -1,6 +1,6 @@
-package com.example.welcomeclient.CallService;
+package com.example.welcomeclient.services;
 
-import com.example.welcomeclient.ResponseDTO;
+import com.example.welcomeclient.models.ResponseDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpMethod;
 import org.springframework.web.client.RestTemplate;
