+++ src/main/java/com/example/welcomeclient/services/RussianRouletteService.java
@@ -1,4 +1,4 @@
-package com.example.welcomeclient.RussianRoulette;
+package com.example.welcomeclient.services;
 
 /**
  * RussianRouletteService should return decision if the bullet was drawn.
